# docs/swagger.yaml
basePath: /api/v1
definitions:
  lib.Page:
    properties:
      first:
        description: indicate first value
        example: true
        type: boolean
      items:
        type: object
      last:
        description: indicate last value
        example: false
        type: boolean
      max_page:
        description: maximum pages for current schema
        example: 9
        type: integer
      page:
        description: current page, start from zero
        example: 0
        type: integer
      size:
        description: size per page, default `10`
        example: 10
        type: integer
      total:
        description: total data without filters
        example: 100
        type: integer
      total_pages:
        description: total pages
        example: 10
        type: integer
      visible:
        description: current length
        example: 10
        type: integer
    type: object
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.CinemaLocation:
    properties:
      additional:
        type: string
      address:
        type: string
      code:
        type: string
      name:
        type: string
    type: object
...
  model.Movie:
    properties:
      additional:
        type: string
      category:
        type: string
      code:
        type: string
      director:
        type: string
      duration:
        type: number
      name:
        type: string
      producer:
        type: string
    type: object
...
  model.ShowSchedule:
    properties:
      additional:
        type: string
      cinema_location:
        $ref: '#/definitions/model.CinemaLocation'
      cinema_location_id:
        type: string
      date:
        format: date
        type: string
      is_start:
        type: boolean
      movie:
        $ref: '#/definitions/model.Movie'
      movie_id:
        type: string
      price:
        type: number
      start_time:
        format: date-time
        type: string
      theater:
        $ref: '#/definitions/model.Theater'
      theater_id:
        type: string
    type: object
  model.ShowScheduleAPI:
    properties:
      cinema_location_id:
        type: string
      date:
        format: date
        type: string
      is_start:
        type: boolean
      movie_id:
        type: string
      price:
        type: number
      start_time:
        format: date-time
        type: string
      theater_id:
        type: string
    type: object
  model.Theater:
    properties:
      additional:
        type: string
      cinema_location_id:
        type: string
      code:
        type: string
      name:
        type: string
    type: object
  model.Ticket:
    properties:
      additional:
        type: string
      date:
        format: date
        type: string
      is_activated:
        type: boolean
      is_printed:
        type: boolean
      location_name:
        type: string
      movie_name:
        type: string
      price:
        type: number
      reference_number:
        type: string
      seat_code:
        type: string
      show_schedule_id:
        type: string
      start_time:
        format: date-time
        type: string
      theater_name:
        type: string
      transaction_id:
        type: string
    type: object
  model.Transaction:
    properties:
      additional:
        type: string
      booking_code:
        type: string
      contact_detail:
        type: string
      contact_name:
        type: string
      invoice_no:
        type: string
      tickets:
        items:
          $ref: '#/definitions/model.Ticket'
        type: array
      total_discount:
        type: number
      total_fee:
        type: number
      total_paid:
        type: number
      total_price:
        type: number
      total_ticket_price:
        type: number
      transaction_status:
        type: string
      user_id:
        format: uuid
        type: string
    type: object
  model.TransactionPayment:
    properties:
      additional:
        type: string
      notes:
        type: string
      paid_amount:
        type: number
      paid_at:
        format: date-time
        type: string
      reference_no:
        type: string
      transaction:
        $ref: '#/definitions/model.Transaction'
      transaction_id:
        format: uuid
        type: string
      type:
        type: string
      via:
        type: string
    type: object
  model.TransactionPaymentAPI:
    properties:
      notes:
        type: string
      paid_amount:
        type: number
      paid_at:
        format: date-time
        type: string
      reference_no:
        type: string
      transaction_id:
        format: uuid
        type: string
      type:
        type: string
      via:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: armadamuhammads@gmail.com
    name: Armada Muhammad Siswanto
  description: API Documentation
  title: MKP Cinema Ticketing
  version: 1.0.0
paths:
...
  /show-schedules:
    get:
      consumes:
      - application/json
      description: List of Show Schedule
      parameters:
      - description: Page number start from zero
        in: query
        name: page
        type: integer
      - description: Size per page, default `0`
        in: query
        name: size
        type: integer
      - description: Sort by field, adding dash (`-`) at the beginning means descending
          and vice versa
        in: query
        name: sort
        type: string
      - description: Select specific fields with comma separated
        in: query
        name: fields
        type: string
      - description: custom filters, see [more details](https://github.com/morkid/paginate#filter-format)
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Show Schedule
          schema:
            allOf:
            - $ref: '#/definitions/lib.Page'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.ShowSchedule'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: List of Show Schedule
      tags:
      - ShowSchedule
    post:
      consumes:
      - application/json
      description: Create new Show Schedule
      parameters:
      - description: Show Schedule data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ShowScheduleAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Show Schedule data
          schema:
            $ref: '#/definitions/model.ShowSchedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new Show Schedule
      tags:
      - ShowSchedule
  /show-schedules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Show Schedule by id
      parameters:
      - description: Show Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Show Schedule by id
      tags:
      - ShowSchedule
    get:
      consumes:
      - application/json
      description: Get a Show Schedule by id
      parameters:
      - description: Show Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Show Schedule data
          schema:
            $ref: '#/definitions/model.ShowSchedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Show Schedule by id
      tags:
      - ShowSchedule
    put:
      consumes:
      - application/json
      description: Update Show Schedule by id
      parameters:
      - description: Show Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Show Schedule data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.ShowScheduleAPI'
      produces:
      - application/json
      responses:
        "200":
          description: Show Schedule data
          schema:
            $ref: '#/definitions/model.ShowSchedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Show Schedule by id
      tags:
      - ShowSchedule
...
  /transactions/{cart_id}:
    post:
      consumes:
      - application/json
      description: Post new Transaction
      parameters:
      - description: Cart ID
        in: path
        name: cart_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Transaction data
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Post new Transaction
      tags:
      - Transaction
  /transactions/{id}:
    get:
      consumes:
      - application/json
      description: Get a Transaction by id
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Transaction by id
      tags:
      - Transaction
  /transactions/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel Unpaid Transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Transaction data
          schema:
            $ref: '#/definitions/model.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Cancel Unpaid Transaction
      tags:
      - Transaction
  /transactions/{id}/payment:
    post:
      consumes:
      - application/json
      description: Pay Transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TransactionPaymentAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction data
          schema:
            $ref: '#/definitions/model.TransactionPayment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Pay Transaction
      tags:
      - Transaction
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  TokenKey:
    in: header
    name: x-Token
    type: apiKey
swagger: "2.0"